bz18130.  Switch over to keeping playlists and items in dicts for sharing.
Track playlists and items on client-side with dicts.  This used to be
lists, e.g. items for a particular playlist was a list.  This was okay
when we didn't need to update stuff after it got created,, but now that
we can update things on the fly it's bad, because we'd need to search
through the whole list (which, could be quite big) for each modification
or deletion we need to track.  So, now just use a dict indexed by the item's
numeric id, and put in appropriate error handling for invalid keys and such.